Small refactor to separate cfdi 3.2
